requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
CacheControl==0.12.11
cachetools==5.2.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
Django==4.0.4
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fcm-django==1.0.12
firebase-admin==5.2.0
geographiclib==1.52
geopy==2.2.0
gevent==21.12.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.3
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
kombu==5.2.4
msgpack==1.0.4
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.31
proto-plus==1.22.0
protobuf==4.21.5
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stripe==4.0.2
toml==0.10.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0